test(build): add tests for AC-6, AC-8, AC-9
New tests:
- api_test: CRUD agents, messages, events, auth rejection (AC-6)
- stale_test: stale agent detection and recovery (AC-8)
- drain_test: drain re-acquire + release (AC-9)

Fixes: make_interval requires int cast, not float.
Evidence: 11 tests pass, 0 failures.

+++ src/models/agent.rs
@@ -148,10 +148,10 @@ pub async fn find_stale_agents(pool: &PgPool, stale_hours: i64) -> Result<Vec<Ag
     let agents = sqlx::query_as::<_, Agent>(
         "SELECT * FROM agents
          WHERE status IN ('awake', 'maintenance')
-           AND wake_started_at < NOW() - make_interval(hours => $1)
+           AND wake_started_at < NOW() - make_interval(hours => $1::int)
         "
     )
-    .bind(stale_hours as f64)
+    .bind(stale_hours as i32)
     .fetch_all(pool)
     .await?;
     Ok(agents)
